requirements.txt
anthropic>=0.40.0
click>=8.0
Pillow>=10.0
pytest>=7.0
